+++ app/src/main/AndroidManifest.xml
@@ -15,19 +15,19 @@
         android:theme="@style/Theme.TopacAdemy_android"
         tools:targetApi="31">
         <activity
-            android:name=".ListActivity"
+            android:name=".car.presentation.ui.ListActivity"
             android:exported="false" />
         <activity
-            android:name=".CalculatorActivity"
+            android:name=".сalculator.presentation.ui.CalculatorActivity"
             android:exported="false" />
         <activity
-            android:name=".WeatherActivity"
+            android:name=".weather.presentation.ui.WeatherActivity"
             android:exported="false" />
         <activity
-            android:name=".SecondActivity"
+            android:name=".second.presentation.ui.SecondActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".main.presentation.ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
