--- app/src/main/java/com/example/topacademy_android/Car.kt
@@ -1,10 +1,0 @@
-package com.example.topacademy_android
-
-data class Car(
-    val brand: String,
-    val model: String,
-    val year: Int,
-    val description: String,
-    val cost: Int,
-    val imageResId: Int
-)
