--- app/src/main/java/com/example/topacademy_android/SecondActivity.kt
@@ -1,61 +1,0 @@
-package com.example.topacademy_android
-
-import android.content.Intent
-import android.os.Bundle
-import android.view.MenuItem
-import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.content.res.AppCompatResources
-import androidx.core.content.ContextCompat
-import com.example.topacademy_android.databinding.ActivitySecondBinding
-
-class SecondActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivitySecondBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = ActivitySecondBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val name = intent.getStringExtra("KEY_NAME")
-        binding.nameView.text = name
-
-        val toolBar = binding.toolBar
-        setSupportActionBar(toolBar)
-
-        // стрелка "назад"
-        supportActionBar?.setDisplayHomeAsUpEnabled(true);
-        supportActionBar?.setDisplayShowHomeEnabled(true);
-
-        // меняем цвет стрелки
-        val arrowDrawable = AppCompatResources.getDrawable(this, androidx.appcompat.R.drawable.abc_ic_ab_back_material)
-        arrowDrawable?.setTint(ContextCompat.getColor(this, R.color.toolbar_icon_color))
-        supportActionBar?.setHomeAsUpIndicator(arrowDrawable)
-
-
-        binding.weatherButton.setOnClickListener{
-            val intent = Intent(this, WeatherActivity::class.java)
-            startActivity(intent)
-        }
-
-        binding.calculatorButton.setOnClickListener{
-            val intent = Intent(this, CalculatorActivity::class.java)
-            startActivity(intent)
-        }
-
-        binding.listButton.setOnClickListener{
-            val intent = Intent(this, ListActivity::class.java)
-            startActivity(intent)
-        }
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        if (item.itemId == android.R.id.home) {
-            finish(); // Возврат на предыдущую Activity
-            return true;
-        }
-        return super.onOptionsItemSelected(item)
-    }
-
-}
