+++ app/src/main/java/com/example/topacademy_android/main/domain/model/ValidationResult.kt
@@ -1,0 +1,3 @@
+package com.example.topacademy_android.main.domain.model
+
+data class ValidationResult(val emailValid: Boolean, val passwordValid: Boolean)
