+++ app/src/main/java/com/example/topacademy_android/main/presentation/event/LoginUiEvent.kt
@@ -1,0 +1,6 @@
+package com.example.topacademy_android.main.presentation.event
+
+sealed class LoginUiEvent {
+    data class ShowToast(val message: String) : LoginUiEvent()
+    data class NavigateToSecond(val username: String) : LoginUiEvent()
+}
