--- app/src/main/java/com/example/topacademy_android/network/DataSeries.kt
@@ -1,9 +1,0 @@
-package com.example.topacademy_android.network
-
-
-data class DataSeries(
-    val date: Int,
-    val weather: String,
-    val wind10m_max: Int,
-    val temp2m: Temperature
-)
