--- app/src/main/java/com/example/topacademy_android/network/Temperature.kt
@@ -1,6 +1,0 @@
-package com.example.topacademy_android.network
-
-data class Temperature (
-    val min : Int,
-    val max : Int
-)
