--- app/src/main/java/com/example/topacademy_android/network/WeatherResponse.kt
@@ -1,7 +1,0 @@
-package com.example.topacademy_android.network
-
-data class WeatherResponse(
-    val product: String,
-    val init: String,
-    val dataseries: List<DataSeries>
-)
