+++ app/src/main/java/com/example/topacademy_android/second/presentation/event/SecondUiEvent.kt
@@ -1,0 +1,7 @@
+package com.example.topacademy_android.second.presentation.event
+
+sealed class SecondUiEvent {
+    data object NavigateToWeather : SecondUiEvent()
+    data object NavigateToCalculator : SecondUiEvent()
+    data object NavigateToCarList : SecondUiEvent()
+}
