+++ app/src/main/java/com/example/topacademy_android/weather/data/remote/api/RetrofitClient.kt
@@ -1,0 +1,25 @@
+package com.example.topacademy_android.weather.data.remote.api
+
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import okhttp3.OkHttpClient
+
+object RetrofitClient {
+
+    private const val BASE_URL = "https://www.7timer.info/"
+
+    val weatherApi: WeatherApi by lazy {
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .client(
+                OkHttpClient.Builder()
+                    .addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
+                    .build()
+            )
+            .build()
+            .create(WeatherApi::class.java)
+    }
+
+}
