+++ app/src/main/java/com/example/topacademy_android/weather/data/remote/api/WeatherApi.kt
@@ -1,0 +1,15 @@
+package com.example.topacademy_android.weather.data.remote.api
+
+import com.example.topacademy_android.weather.domain.model.WeatherResponse
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface WeatherApi {
+    @GET("bin/api.pl")
+    suspend  fun getWeatherForecast(
+        @Query("lon") longitude: Double,
+        @Query("lat") latitude: Double,
+        @Query("product") product: String = "civillight",
+        @Query("output") output: String = "json"
+    ): WeatherResponse
+}
