+++ app/src/main/java/com/example/topacademy_android/weather/domain/model/DataSeries.kt
@@ -1,0 +1,9 @@
+package com.example.topacademy_android.weather.domain.model
+
+
+data class DataSeries(
+    val date: Int,
+    val weather: String,
+    val wind10m_max: Int,
+    val temp2m: Temperature
+)
