+++ app/src/main/java/com/example/topacademy_android/weather/domain/model/Temperature.kt
@@ -1,0 +1,6 @@
+package com.example.topacademy_android.weather.domain.model
+
+data class Temperature (
+    val min : Int,
+    val max : Int
+)
