+++ app/src/main/java/com/example/topacademy_android/weather/domain/model/WeatherResponse.kt
@@ -1,0 +1,7 @@
+package com.example.topacademy_android.weather.domain.model
+
+data class WeatherResponse(
+    val product: String,
+    val init: String,
+    val dataseries: List<DataSeries>
+)
