+++ app/src/main/java/com/example/topacademy_android/сalculator/presentation/viewmodel/CalculatorViewModel.kt
@@ -1,0 +1,29 @@
+package com.example.topacademy_android.сalculator.presentation.viewmodel
+
+import androidx.lifecycle.ViewModel
+import com.example.topacademy_android.сalculator.domain.use_case.CalculatorUseCase
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+
+class CalculatorViewModel(private val useCase: CalculatorUseCase) : ViewModel(){
+
+    private val _expression = MutableStateFlow("")
+    val expression: StateFlow<String> = _expression
+
+    fun appendSymbol(symbol: String) {
+        _expression.value += symbol
+    }
+
+    fun clear() {
+        _expression.value = ""
+    }
+
+    fun evaluate() {
+        _expression.value = useCase.evaluateExpression(_expression.value)
+    }
+
+    fun backspace() {
+        _expression.value = _expression.value.dropLast(1)
+    }
+
+}
