+++ app/src/main/res/layout/activity_calculator.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CalculatorActivity">
+    tools:context=".сalculator.presentation.ui.CalculatorActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/tool_bar"
@@ -325,7 +325,7 @@
         android:text="@string/coma"
 
         app:layout_constraintStart_toEndOf="@+id/btn0"
-        app:layout_constraintEnd_toStartOf="@+id/plus_or_minus"
+        app:layout_constraintEnd_toStartOf="@+id/plusMinus"
         app:layout_constraintTop_toBottomOf="@+id/btn2"
 
         android:layout_marginTop="@dimen/button_calc_margin_top"
@@ -333,7 +333,7 @@
         android:layout_marginStart="@dimen/button_calc_margin_left_right"/>
 
     <com.google.android.material.button.MaterialButton
-        android:id="@+id/plus_or_minus"
+        android:id="@+id/plusMinus"
         style="@style/button_style"
         android:layout_height="@dimen/button_calc_height"
         android:layout_width="wrap_content"
@@ -354,7 +354,7 @@
         android:layout_width="wrap_content"
         android:text="@string/equally"
 
-        app:layout_constraintStart_toEndOf="@+id/plus_or_minus"
+        app:layout_constraintStart_toEndOf="@+id/plusMinus"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/addition"
