+++ app/src/main/res/layout/activity_list.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ListActivity">
+    tools:context=".car.presentation.ui.ListActivity">
 
     <androidx.appcompat.widget.Toolbar
         style="@style/tool_bar_style"
