+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     style="@style/layout_style"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".main.presentation.ui.MainActivity">
 
 
     <com.google.android.material.textfield.TextInputLayout
